pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cogen-align"
version = "0.1.0"
description = "Contrastive warmup + generative alignment for SpeechLM (CoGen-Align)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.1",
  "transformers>=4.40",
  "peft>=0.11",
  "pyyaml>=6.0",
  "numpy>=1.24",
  "safetensors>=0.4",
]

[project.optional-dependencies]
train = ["wandb>=0.16", "jiwer>=3.0"]
dev = ["pytest>=7.0", "ruff>=0.4"]

[tool.setuptools.packages.find]
where = ["src"]
